:truck: move from Slider/index code to Slider file.
