+++ src/components/common/Slider/index.tsx
@@ -1,58 +1,14 @@
-import { ReactElement } from 'react';
-import { css } from 'styled-components';
-
-import { SliderOptions } from '@components/common/Slider/context/SliderInfoContext';
-import SliderProvider from '@components/common/Slider/context/SliderProvider';
 import NextButton from '@components/common/Slider/NextButton';
 import PrevButton from '@components/common/Slider/PrevButton';
 import Slide from '@components/common/Slider/Slide';
 import SlideList from '@components/common/Slider/SlideList';
+import SliderRoot from '@components/common/Slider/Slider';
 
-type SliderProps = {
-  options?: Partial<SliderOptions>;
-  children: ReactElement | ReactElement[];
-};
-
-const defaultOptions = {
-  slidesToShow: 1,
-  slidesToScroll: 1,
-  slidesMargin: '0px',
-  speed: 1000,
-  initialSlide: 0,
-};
-
-const Slider = ({ options = defaultOptions, children }: SliderProps) => {
-  const sliderOptions = { ...defaultOptions, ...options };
-
-  const getSlideLength = () => {
-    const Children = Array.isArray(children) ? children : [children];
-    const slideList = Children.find((child) => {
-      if (typeof child.type === 'string') return false;
-      return child?.type.name === 'SlideList';
-    });
-    const SlideLength = slideList ? slideList.props.children.length : 0;
-
-    return SlideLength;
-  };
-
-  return (
-    <SliderProvider options={sliderOptions} SlideLength={getSlideLength()}>
-      <div
-        css={css`
-          position: relative;
-          width: 100%;
-          overflow: hidden;
-        `}
-      >
-        {children}
-      </div>
-    </SliderProvider>
-  );
-};
-
-Slider.List = SlideList;
-Slider.Item = Slide;
-Slider.PrevButton = PrevButton;
-Slider.NextButton = NextButton;
+const Slider = Object.assign(SliderRoot, {
+  List: SlideList,
+  Item: Slide,
+  PrevButton,
+  NextButton,
+});
 
 export default Slider;
